add camelcase|camelize, pluralize, underscore, upcase|uppercase filters
These use the inflection library from github/phoffer/inflector.cr so that has
been added as a dependency. The case changing filters were added to what was
filters/downcase.cr and the file renamed to filters/case.cr as a collection
of filters that alter the case of the input.